Move some header blob stuff into an internal module.
