+++ test/Data/RpmDbHeaderBlobSpec.hs
@@ -6,7 +6,7 @@ import Data.Either (fromRight)
 import Data.Int (Int32)
 import Data.List (isSuffixOf)
 import Data.List.NonEmpty qualified as NonEmpty
-import Data.Rpm.DbHeaderBlob (
+import Data.Rpm.DbHeaderBlob.Internal (
   EntryInfo (..),
   HeaderBlob (..),
   IndexEntry (..),
